# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Portfolio Strategist API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserMarketStatus =
  (typeof UserMarketStatus)[keyof typeof UserMarketStatus];

export const UserMarketStatus = {
  Open: "Open",
  Closed: "Closed",
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  avatarUrl?: string | null;
  marketStatus: UserMarketStatus;
  marketTime: string;
}

export type InvestmentGoalsRiskTolerance =
  (typeof InvestmentGoalsRiskTolerance)[keyof typeof InvestmentGoalsRiskTolerance];

export const InvestmentGoalsRiskTolerance = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface InvestmentGoals {
  totalCapital: number;
  targetReturnPct: number;
  timeHorizonMonths: number;
  riskTolerance: InvestmentGoalsRiskTolerance;
  goalProgressPct: number;
  goalProgressNote?: string;
}

export type UpdateGoalsRequestRiskTolerance =
  (typeof UpdateGoalsRequestRiskTolerance)[keyof typeof UpdateGoalsRequestRiskTolerance];

export const UpdateGoalsRequestRiskTolerance = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface UpdateGoalsRequest {
  totalCapital: number;
  targetReturnPct: number;
  timeHorizonMonths: number;
  riskTolerance: UpdateGoalsRequestRiskTolerance;
}

export interface AllocationSlice {
  assetClass: string;
  percentage: number;
}

export type StrategyRiskLevel =
  (typeof StrategyRiskLevel)[keyof typeof StrategyRiskLevel];

export const StrategyRiskLevel = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface Strategy {
  strategyType: string;
  lastGenerated: string;
  riskLevel: StrategyRiskLevel;
  allocation: AllocationSlice[];
  keyRules: string[];
}

export interface Portfolio {
  totalValue: number;
  totalProfitLoss: number;
  totalProfitLossPct: number;
  change24h: number;
  change24hPct: number;
}

export type AllocationRowStatus =
  (typeof AllocationRowStatus)[keyof typeof AllocationRowStatus];

export const AllocationRowStatus = {
  In_Range: "In Range",
  Slightly_High: "Slightly High",
  Slightly_Low: "Slightly Low",
  Out_of_Range: "Out of Range",
} as const;

export interface AllocationRow {
  assetClass: string;
  currentPct: number;
  targetPct: number;
  differencePct: number;
  status: AllocationRowStatus;
}

export interface Holding {
  id: string;
  symbol: string;
  name: string;
  assetClass: string;
  quantity: number;
  price: number;
  value: number;
  change24hPct: number;
}

export interface PerformancePoint {
  timestamp: string;
  portfolio: number;
  benchmark: number;
}

export interface PerformanceSeries {
  range: string;
  points: PerformancePoint[];
}

export type TradeSuggestionSide =
  (typeof TradeSuggestionSide)[keyof typeof TradeSuggestionSide];

export const TradeSuggestionSide = {
  Buy: "Buy",
  Sell: "Sell",
  Hold: "Hold",
} as const;

export type TradeSuggestionStatus =
  (typeof TradeSuggestionStatus)[keyof typeof TradeSuggestionStatus];

export const TradeSuggestionStatus = {
  Open: "Open",
  Applied: "Applied",
  Dismissed: "Dismissed",
} as const;

export interface TradeSuggestion {
  id: string;
  symbol: string;
  pair: string;
  side: TradeSuggestionSide;
  status: TradeSuggestionStatus;
  entryRangeLow: number;
  entryRangeHigh: number;
  target: number;
  stopLoss: number;
  positionSize?: string;
  suggestedAction?: string;
  reasoning?: string;
  riskWarning: string;
  createdAt: string;
  summary: string;
}

export type RebalancingActionActionType =
  (typeof RebalancingActionActionType)[keyof typeof RebalancingActionActionType];

export const RebalancingActionActionType = {
  Buy: "Buy",
  Sell: "Sell",
} as const;

export interface RebalancingAction {
  id: string;
  asset: string;
  actionType: RebalancingActionActionType;
  amount: number;
}

export interface ApplyRebalancingResult {
  appliedCount: number;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  Buy: "Buy",
  Sell: "Sell",
  Deposit: "Deposit",
  Withdraw: "Withdraw",
  Dividend: "Dividend",
  Rebalance: "Rebalance",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  Completed: "Completed",
  Pending: "Pending",
  Failed: "Failed",
} as const;

export interface Transaction {
  id: string;
  occurredAt: string;
  type: TransactionType;
  asset: string;
  amount: number;
  value: number;
  status: TransactionStatus;
  note?: string;
}

export type RiskAlertSeverity =
  (typeof RiskAlertSeverity)[keyof typeof RiskAlertSeverity];

export const RiskAlertSeverity = {
  Info: "Info",
  Warning: "Warning",
  Critical: "Critical",
} as const;

export interface RiskAlert {
  id: string;
  severity: RiskAlertSeverity;
  title: string;
  message: string;
  createdAt: string;
  dismissed: boolean;
}

export type AssistantMessageRole =
  (typeof AssistantMessageRole)[keyof typeof AssistantMessageRole];

export const AssistantMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface AssistantMessage {
  id: string;
  role: AssistantMessageRole;
  content: string;
  createdAt: string;
}

export interface SendAssistantRequest {
  content: string;
}

export interface RunCommandRequest {
  prompt: string;
}

export interface CommandResult {
  reply: string;
}

export interface DashboardSummary {
  portfolio: Portfolio;
  goals: InvestmentGoals;
  strategy: Strategy;
  alertsCount: number;
  healthScore: number;
  healthLabel: string;
  healthNote?: string;
}

export type GetPerformanceSeriesParams = {
  range?: GetPerformanceSeriesRange;
};

export type GetPerformanceSeriesRange =
  (typeof GetPerformanceSeriesRange)[keyof typeof GetPerformanceSeriesRange];

export const GetPerformanceSeriesRange = {
  "1D": "1D",
  "7D": "7D",
  "1M": "1M",
  "3M": "3M",
  "1Y": "1Y",
  ALL: "ALL",
} as const;

export type ListTransactionsParams = {
  limit?: number;
};

export type SendAssistantMessage200 = {
  userMessage: AssistantMessage;
  assistantMessage: AssistantMessage;
};