# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Portfolio Strategist API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades/{id}/apply:
    post:
      operationId: applyTradeSuggestion
      tags: [trades]
      summary: Apply (accept) a trade suggestion
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated suggestion
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TradeSuggestion" }
...
  /alerts/{id}/dismiss:
    post:
      operationId: dismissAlert
      tags: [alerts]
      summary: Dismiss an alert
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Dismissed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RiskAlert" }
  /assistant/messages:
    get:
      operationId: listAssistantMessages
      tags: [assistant]
      summary: AI Trade Assistant chat history
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AssistantMessage" }
    post:
      operationId: sendAssistantMessage
      tags: [assistant]
      summary: Send a message to the AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendAssistantRequest" }
      responses:
        "200":
          description: New assistant message pair (user + reply)
          content:
            application/json:
              schema:
                type: object
                required: [userMessage, assistantMessage]
                properties:
                  userMessage: { $ref: "#/components/schemas/AssistantMessage" }
                  assistantMessage: { $ref: "#/components/schemas/AssistantMessage" }
  /command:
    post:
      operationId: runCommand
      tags: [assistant]
      summary: Run a command/question via the AI Command Center
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RunCommandRequest" }
      responses:
        "200":
          description: Command response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommandResult" }
components:
  schemas:
...
    TradeSuggestion:
      type: object
      required: [id, symbol, pair, side, status, entryRangeLow, entryRangeHigh, target, stopLoss, riskWarning, createdAt, summary]
      properties:
        id: { type: string }
        symbol: { type: string }
        pair: { type: string }
        side: { type: string, enum: [Buy, Sell, Hold] }
        status: { type: string, enum: [Open, Applied, Dismissed] }
        entryRangeLow: { type: number }
        entryRangeHigh: { type: number }
        target: { type: number }
        stopLoss: { type: number }
        positionSize: { type: string }
        suggestedAction: { type: string }
        reasoning: { type: string }
        riskWarning: { type: string }
        createdAt: { type: string }
        summary: { type: string }
...
    RiskAlert:
      type: object
      required: [id, severity, title, message, createdAt, dismissed]
      properties:
        id: { type: string }
        severity: { type: string, enum: [Info, Warning, Critical] }
        title: { type: string }
        message: { type: string }
        createdAt: { type: string }
        dismissed: { type: boolean }
    AssistantMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id: { type: string }
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
        createdAt: { type: string }
    SendAssistantRequest:
      type: object
      required: [content]
      properties:
        content: { type: string }
    RunCommandRequest:
      type: object
      required: [prompt]
      properties:
        prompt: { type: string }
    CommandResult:
      type: object
      required: [reply]
      properties:
        reply: { type: string }
...